% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cokelet_mld.R
\name{calculate_cokelet_mld}
\alias{calculate_cokelet_mld}
\title{Calculate mixed layer depth using a threshold method}
\usage{
calculate_cokelet_mld(rho, z, ref.depths, totdepth, threshold = 0.1)
}
\arguments{
\item{rho}{Numeric vector of densities}

\item{z}{Numeric vector of depths. Depths are positive.}

\item{ref.depths}{Depths to use for calculating upper water column density}

\item{totdepth}{Maximum depth sampled by the cast}

\item{threshold}{Density threshold for changes.}
}
\description{
Following Danielsson et al. 2011 and Cokelet 2016.
}
