% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density_diff.R
\name{calculate_density_diff}
\alias{calculate_density_diff}
\title{Calculate density difference}
\usage{
calculate_density_diff(
  lower.depth = 50,
  upper.depth = 5,
  ctd.dat = paste0(seantools::root.dir(),
    "Thesis/Chapter 1/data/CTD/output/2019-04-11_ebs_shelf_ctd_2008_2018.rds")
)
}
\arguments{
\item{lower.depth}{Lower depth for calculating density difference. Default = 50 m}

\item{upper.depth}{Upper depth for calculating density difference. Default = 5 m averages density in the upper 5 m.}

\item{ctd.dat}{File path to CTD data.}
}
\description{
Import and manipulate CTD data to only get cast data from good quality bottom-trawl survey hauls, then calculate density difference between Z <=5 m and z = 50
}
