% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tsrho_by_layer.R
\name{calculate_tsrho_by_layer}
\alias{calculate_tsrho_by_layer}
\title{Calculate temperature, salinity, and density for each layer}
\usage{
calculate_tsrho_by_layer(t, s, rho, z, mld, mld.buffer)
}
\arguments{
\item{t}{temperature vector}

\item{s}{Salinity vector}

\item{rho}{Density vector}

\item{z}{Depth vector}

\item{mld}{Mixed layer depth}
}
\description{
Calculate temperature, salinity, and density for each layer
}
