% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_wrapper.R
\name{cast_wrapper}
\alias{cast_wrapper}
\title{Applies a function to all casts in a data frame}
\usage{
cast_wrapper(x, id.col, FUN, min.rows = 4, silent = TRUE, ...)
}
\description{
Applies a function to all casts in a data frame
}
