% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumR.R
\name{illumR}
\alias{illumR}
\title{Vectorized version of Janiczek and DeYoung (1987) for R}
\usage{
illumR(
  IY,
  IM,
  ID,
  LO,
  FINIT,
  ZZ,
  SK,
  HR,
  full.output = FALSE,
  match.mk = FALSE
)
}
\arguments{
\item{IY}{Input year.}

\item{IM}{Input month}

\item{ID}{Input day}

\item{LO}{Longitude}

\item{FINIT}{LATITUDE}

\item{ZZ}{Which timezone (0 = GMT, 1 = standard zone time, 2 = local mean time)}

\item{SK}{Sky condition. (1 = Sun/moon visible, sky <70% overcast, 2 = sun/moon obscured by thin clouds, 3 = sun/moon obscured by average clouds, 10 = sun/moon obscured by dark stratus clouds (rare))}

\item{HR}{Time based on a 24 hour clock, as a numeric vector (e.g. enter 1330 for 13:30, 820 for 8:20)}

\item{full.output}{If true, returns sun and moon separately.}

\item{match.mk}{In testing.}
}
\value{
Illuminance (lux) at the Earth's surface.
}
\description{
Vectorized version of Janiczek and DeYoung (1987) for R
}
\references{
Janiczek, P. M., and DeYoung, J.A. 1987. Computer programs for sun and moon illuminance with contingent tables and diagrams. US Nav. Observ. Circ. 171.
}
