% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_2.R
\name{loocv_2}
\alias{loocv_2}
\title{Perform leave-one-out cross-validation and spatial interpolation on EBS data}
\usage{
loocv_2(
  dat,
  var.col,
  lat.col,
  lon.col,
  in.proj = "+proj=longlat +datum=NAD83",
 
    interp.proj = "+proj=aea +lat_1=55 +lat_2=65 +lat_0=50 +lon_0=-154 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs",
  scale.vars = FALSE,
  center = TRUE,
  scale = TRUE,
  nm = Inf,
  pre = NA,
  ...
)
}
\arguments{
\item{dat}{Data frame that contains latitude, longitude, and a variable to be interpolated.}

\item{var.col}{Name of the variable column that is to be interpolated}

\item{lat.col}{Name of the latitude column}

\item{lon.col}{Name of the longitude column}

\item{in.proj}{Projection for the input data}

\item{interp.proj}{Projection for the interpolation/output}

\item{scale.vars}{Should variable be scaled?}

\item{center}{Passed to \code{scale} if scale.vars == TRUE}

\item{scale}{Passed to \code{scale}  if scale.vars == TRUE}

\item{nm}{Maximum number of neighboring stations to use for interpolation.}

\item{pre}{}
}
\description{
Sean K. Rohan <sean.rohan@noaa.gov>
Last update: February 28, 2019
Spatial interpolation with interpolation method selection based on leave-one-out cross validation
}
