% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all.R
\name{process_all}
\alias{process_all}
\title{Wrapper for filter_stepwise and calculate_attenuation}
\usage{
process_all(
  dir.structure,
  time.buffer = 20,
  cast.dir = "Downcast",
  agg.fun = median,
  binsize = 2,
  bin.gap = 6,
  kz.binsize = 0.5,
  silent = TRUE,
  ...
)
}
\description{
This function is designed to work with the RACE data structure to subset. all of the Mk9 data obtained during upscasts and downcasts and runs functions convert_light, filter_stepwise, and calculate_attenuation on the data. The directory structure needs to be input for the function to work.
}
