% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ctd_fns.R
\name{run_ctd_fns}
\alias{run_ctd_fns}
\title{Wrapper function for calculating mld, density difference between upper and lower water column, and pycnocline depth}
\usage{
run_ctd_fns(dat)
}
\description{
Wrapper function for calculating mld, density difference between upper and lower water column, and pycnocline depth
}
