% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_adjustments.R
\name{time_adjustments}
\alias{time_adjustments}
\title{MCorrect tag time}
\usage{
time_adjustments(light.data, cast.data)
}
\description{
Make adjustments to correct inconsistencies between tag time and survey time.
}
