% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{vertical_profiles}
\alias{vertical_profiles}
\title{Subsets light measurements from upcasts/downcasts}
\usage{
vertical_profiles(light.data, cast.data, ...)
}
\description{
Assigns light measurements to upcast or downcast based on downcast start time and downcast end time.
}
